requirements.txt
boto==2.48.0
brotlipy==0.7.0
cachetools==2.0.1
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.4.2
Django==2.0
django-ajax-selects==1.7.0
django-allauth==0.34.0
django-crispy-forms==1.7.0
django-oauth-toolkit==1.0.0
django-storages==1.6.5
django-webpack-loader==0.5.0
djangorestframework==3.7.7
flake8==3.5.0
google-api-core==0.1.4
google-auth==1.4.1
google-cloud-core==0.28.0
google-cloud-storage==1.7.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
gunicorn==19.7.1
idna==2.6
mccabe==0.6.1
oauthlib==2.0.6
Pillow==5.0.0
protobuf==3.5.1
psycopg2==2.7.3.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
python-dateutil==2.6.1
python-mimeparse==1.6.0
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
six==1.11.0
urllib3==1.22
whitenoise==3.3.1
